Track Go module data
https://github.com/golang/go/wiki/Modules#should-i-commit-my-gosum-file-as-well-as-my-gomod-file
Which allows to simply run "go build" after cloning the repo.

Theoretically works with the latest versions, but go-runewidth not updated to 0.0.14. Calls in a new dep with an new indirect dep, which would need to stay at 0.0.2

Additionally adjust the import statement for the submodules, as without adjustments "go build" would fail because of not found packages in GOROOT